+++ packages/vite/src/node/plugins/esbuild.ts
@@ -11,7 +11,7 @@ import type { RawSourceMap } from '@jridgewell/remapping'
 import type { InternalModuleFormat, SourceMap } from 'rollup'
 import type { TSConfckParseResult } from 'tsconfck'
 import { TSConfckCache, TSConfckParseError, parse } from 'tsconfck'
-import type { FSWatcher } from 'dep-types/chokidar'
+import type { FSWatcher } from '#dep-types/chokidar'
 import {
   combineSourcemaps,
   createDebugger,
@@ -29,7 +29,7 @@ const debug = createDebugger('vite:esbuild')
 // IIFE content looks like `var MyLib = function() {`.
 // Spaces are removed and parameters are mangled when minified
 const IIFE_BEGIN_RE =
-  /(?:const|var)\s+\S+\s*=\s*function\([^()]*\)\s*\{\s*"use strict";/
+  /(?:const|var)\s+\S+\s*=\s*\(?function\([^()]*\)\s*\{\s*"use strict";/
 
 const validExtensionRE = /\.\w+$/
 const jsxExtensionsRE = /\.(?:j|t)sx\b/
@@ -249,7 +249,6 @@ export function esbuildPlugin(config: ResolvedConfig): Plugin {
   // and for build as the final optimization is in `buildEsbuildPlugin`
   const transformOptions: TransformOptions = {
     target: 'esnext',
-    charset: 'utf8',
     ...esbuildTransformOptions,
     minify: false,
     minifyIdentifiers: false,
@@ -403,7 +402,6 @@ export function resolveEsbuildTranspileOptions(
   const esbuildOptions = config.esbuild || {}
 
   const options: TransformOptions = {
-    charset: 'utf8',
     ...esbuildOptions,
     loader: 'js',
     target: target || undefined,