+++ packages/vite/src/node/server/ws.ts
@@ -9,10 +9,10 @@ import crypto from 'node:crypto'
 import colors from 'picocolors'
 import type { WebSocket as WebSocketRaw } from 'ws'
 import { WebSocketServer as WebSocketServerRaw_ } from 'ws'
-import type { WebSocket as WebSocketTypes } from 'dep-types/ws'
-import type { ErrorPayload, HotPayload } from 'types/hmrPayload'
-import type { InferCustomEventPayload } from 'types/customEvent'
 import { isHostAllowed } from 'host-validation-middleware'
+import type { WebSocket as WebSocketTypes } from '#dep-types/ws'
+import type { ErrorPayload, HotPayload } from '#types/hmrPayload'
+import type { InferCustomEventPayload } from '#types/customEvent'
 import type { ResolvedConfig } from '..'
 import { isObject } from '../utils'
 import type { NormalizedHotChannel, NormalizedHotChannelClient } from './hmr'
@@ -35,7 +35,7 @@ export type WebSocketCustomListener<T> = (
   invoke?: 'send' | `send:${string}`,
 ) => void
 
-export const isWebSocketServer = Symbol('isWebSocketServer')
+export const isWebSocketServer: unique symbol = Symbol('isWebSocketServer')
 
 export interface WebSocketServer extends NormalizedHotChannel {
   /**